Fix mismatch between aosp and internal.
Somehow, a relation chain of cherry-picked changes submitted
in the wrong order resulting in one change being applied too
late. This resulted in a small change being lost. Luckily, it
was only this one single change, so I'm applying that change
manually.

I verified that aosp master and rvc-dev-plus-aosp are fine and
the changes landed in the correct order.

Test: Boots on a bonito and scudo unit tests pass.
Change-Id: I2ec88ddc4d9acdbab1c2e8191383bc8262d29843