scudo: Remove ANDROID_EXPERIMENTAL_MTE macro.
Kernel support for MTE has been released in Linux 5.10. This means
that it is a stable API and we no longer need to make the support
conditional on a macro. We do need to provide conditional definitions
of the new macros though in order to avoid a dependency on new
kernel headers.

Differential Revision: https://reviews.llvm.org/D93513

GitOrigin-RevId: dfa40840e0e2fa094c5d3f441affe0785cdc8d09
Change-Id: I371ea42add48d89522eb22c87b822cb3695a4dd4