[scudo] Use stdint types for internal scudo types
`scudo::uptr` was defined as an `unsigned long` on 32-b platform,
while a `uintptr_t` is usually defined as an `unsigned int`.
This worked, this was not consistent, particularly with regard to
format string specifiers.

As suggested by Vitaly, since we are including `stdint.h`, define
the internal `scudo` integer types to those.

Differential Revision: https://reviews.llvm.org/D108089

GitOrigin-RevId: 5fc841d8a278ea16bae457deba35d0db6b716dd6
Change-Id: I449369fdb0a5f04c717dad75b45a4e9996f4d5ec