[scudo] Use stdint types for internal types (redo)
This is a redo of D108089 that broke some 32-bit builds.

`scudo::uptr` was defined as an `unsigned long` on 32-b platform,
while a `uintptr_t` is usually defined as an `unsigned int`.
This worked, this was not consistent, particularly with regard to
format string specifiers.

As suggested by Vitaly, since we are including `stdint.h`, define
the internal scudo integer types to those.

Differential Revision: https://reviews.llvm.org/D108152

GitOrigin-RevId: b2aaafb8377ac1ab081e7c0d3ba92ee5eb4de07c
Change-Id: Id16c7b4652b873a759ac2f1523e59d3ab157a298